requirements.txt
django
djangorestframework
pyjwt
django-cors-headers
